requirements.txt
# Core Framework
streamlit==1.31.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.27

# AI / LLM
google-generativeai==0.4.0

# Data Processing
pandas==2.2.0
numpy==1.26.4

# Visualization
plotly==5.19.0

# Security & Validation
pydantic==2.6.1
python-dotenv==1.0.1

# Utilities
python-dateutil==2.8.2
